+++ src/main/java/org/rundeck/client/tool/Main.java
@@ -153,7 +153,8 @@ public class Main {
                                         new Tokens(rd),
                                         new Nodes(rd),
                                         new Users(rd),
-                                        new Something()
+                                        new Something(),
+                                        new Retry(rd)
 
                                 )
                                 .bannerResource("rd-banner.txt")